+++ CHANGELOG.md
@@ -9,6 +9,7 @@ for further information about branching and tagging conventions.
 
 ## [Unreleased]
 #### Added
+- Display fees in invest form
 - "initSync" method to Api class
 
 #### Changed
@@ -17,6 +18,18 @@ for further information about branching and tagging conventions.
 
 #### Removed
 - Unused methods from `DocumentContainer` class
+
+## [1.6.0-x.2] - 2019-04-29
+#### Added
+- Re-render chart animation
+
+#### Changed
+- Now opening pre-issuance details link in new tab
+
+### Experimental features changes
+#### Fixed
+- Deposit fiat drawer displaying on REIT
+>>>>>>> master
 
 ## [1.6.0-x.1] - 2019-04-26
 #### Added
@@ -37,7 +50,7 @@ for further information about branching and tagging conventions.
 ### "Under the hood" changes
 #### Changed
 - Now using @tokend/js-sdk@1.6.0-x.0
-- Now using "app__button-*" classes instead of button mixins
+- Now using "app__button-..." classes instead of button mixins
 
 ### Experimental features changes
 #### Changed
@@ -391,7 +404,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.6.0-x.1...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.6.0-x.2...HEAD
+[1.6.0-x.2]: https://github.com/tokend/web-client/compare/1.6.0-x.1...1.6.0-x.2
 [1.6.0-x.1]: https://github.com/tokend/web-client/compare/1.6.0-x.0...1.6.0-x.1
 [1.6.0-x.0]: https://github.com/tokend/web-client/compare/1.5.0...1.6.0-x.0
 [1.5.0]: https://github.com/tokend/web-client/compare/1.5.0-rc.3...1.5.0